List and card

+++ src/main/java/com/sparta/springtrello/domain/card/service/CardService.java
@@ -2,15 +2,21 @@ package com.sparta.springtrello.domain.card.service;
 
 import java.util.List;
 
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.sparta.springtrello.domain.board.entitiy.Board;
+import com.sparta.springtrello.domain.board.repository.BoardRepository;
 import com.sparta.springtrello.domain.card.dto.CardRequestDto;
 import com.sparta.springtrello.domain.card.entity.Card;
 import com.sparta.springtrello.domain.card.repository.CardRepository;
 import com.sparta.springtrello.domain.common.dto.AuthUser;
 import com.sparta.springtrello.domain.list.entity.BoardList;
 import com.sparta.springtrello.domain.list.repository.ListRepository;
+import com.sparta.springtrello.domain.user.entity.UserWorkspace;
+import com.sparta.springtrello.domain.user.enums.WorkspaceUserRole;
+import com.sparta.springtrello.domain.user.repository.UserWorkspaceRepository;
 
 import lombok.RequiredArgsConstructor;
 
@@ -21,6 +27,8 @@ public class CardService {
 
     private final CardRepository cardRepository;
     private final ListRepository listRepository;
+    private final BoardRepository boardRepository;
+    private final UserWorkspaceRepository userWorkspaceRepository;
 
     // 카드 생성
     @Transactional
@@ -30,6 +38,23 @@ public class CardService {
                 listRepository
                         .findById(cardRequestDto.getListId())
                         .orElseThrow(() -> new IllegalArgumentException("해당 리스트를 찾을 수 없습니다."));
+
+        Board board =
+                boardRepository
+                        .findById(boardList.getBoardId())
+                        .orElseThrow(
+                                () -> new IllegalArgumentException("해당 리스트가 속한 보드를 찾을 수 없습니다."));
+
+        UserWorkspace userWorkspace =
+                userWorkspaceRepository
+                        .findByUserIdAndWorkspaceId(authUser.getId(), board.getWorkspaceId())
+                        .orElseThrow(
+                                () -> new IllegalArgumentException("해당 워크스페이스에 대한 접근 권한이 없습니다."));
+
+        if (userWorkspace.getWorkspaceUserRole() == WorkspaceUserRole.READ_ONLY) {
+            throw new AccessDeniedException("읽기 전용 권한을 가진 유저는 카드를 생성할 수 없습니다.");
+        }
+
         // 카드 생성
         Card card =
                 new Card(
@@ -39,6 +64,7 @@ public class CardService {
                         cardRequestDto.getDueDate(),
                         cardRequestDto.getManagerId(),
                         cardRequestDto.getListId());
+
         return cardRepository.save(card);
     }
 
@@ -57,16 +83,30 @@ public class CardService {
     @Transactional
     public Card updateCard(Long cardId, AuthUser authUser, CardRequestDto cardRequestDto) {
 
-        // 카드 존재 여부 확인
         Card card =
                 cardRepository
                         .findById(cardId)
                         .orElseThrow(() -> new IllegalArgumentException("해당 카드를 찾을 수 없습니다."));
-
         BoardList boardList =
                 listRepository
                         .findById(cardRequestDto.getListId())
                         .orElseThrow(() -> new IllegalArgumentException("해당 리스트를 찾을 수 없습니다."));
+
+        Board board =
+                boardRepository
+                        .findById(boardList.getBoardId())
+                        .orElseThrow(
+                                () -> new IllegalArgumentException("해당 리스트가 속한 보드를 찾을 수 없습니다."));
+
+        UserWorkspace userWorkspace =
+                userWorkspaceRepository
+                        .findByUserIdAndWorkspaceId(authUser.getId(), board.getWorkspaceId())
+                        .orElseThrow(
+                                () -> new IllegalArgumentException("해당 워크스페이스에 대한 접근 권한이 없습니다."));
+
+        if (userWorkspace.getWorkspaceUserRole() == WorkspaceUserRole.READ_ONLY) {
+            throw new AccessDeniedException("읽기 전용 권한을 가진 유저는 카드를 생성할 수 없습니다.");
+        }
 
         // 카드 정보 수정
         card.updateCard(
@@ -85,11 +125,31 @@ public class CardService {
     @Transactional
     public void deleteCard(Long cardId, AuthUser authUser) {
 
-        // 카드 존재 여부 확인
         Card card =
                 cardRepository
                         .findById(cardId)
                         .orElseThrow(() -> new IllegalArgumentException("해당 카드를 찾을 수 없습니다."));
+        BoardList boardList =
+                listRepository
+                        .findById(card.getListId())
+                        .orElseThrow(
+                                () -> new IllegalArgumentException("해당 카드가 속한 리스트를 찾을 수 없습니다."));
+
+        Board board =
+                boardRepository
+                        .findById(boardList.getBoardId())
+                        .orElseThrow(
+                                () -> new IllegalArgumentException("해당 리스트가 속한 보드를 찾을 수 없습니다."));
+
+        UserWorkspace userWorkspace =
+                userWorkspaceRepository
+                        .findByUserIdAndWorkspaceId(authUser.getId(), board.getWorkspaceId())
+                        .orElseThrow(
+                                () -> new IllegalArgumentException("해당 워크스페이스에 대한 접근 권한이 없습니다."));
+
+        if (userWorkspace.getWorkspaceUserRole() == WorkspaceUserRole.READ_ONLY) {
+            throw new AccessDeniedException("읽기 전용 권한을 가진 유저는 카드를 삭제할 수 없습니다.");
+        }
 
         // 카드 삭제
         cardRepository.delete(card);
